Move node normalization to a mixin
